requirements.txt
alembic==1.13.2
blinker==1.8.2
cffi==1.17.1
click==8.1.7
cryptography==43.0.1
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
greenlet==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.9.0
PyYAML==6.0.2
SQLAlchemy==2.0.34
typing_extensions==4.12.2
uuid==1.30
Werkzeug==3.0.4
